# tradit-edit/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  brand: string;
  name: string;
  category: string;
  sizeRange: string;
  color: string;
  price: number;
  originalPrice: number;
  imageUrl: string;
  /** @nullable */
  badge?: string | null;
  description: string;
  availableSizes: string[];
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  trade_in: 'trade_in',
  purchase: 'purchase',
  impact: 'impact',
} as const;

export interface Activity {
  id: string;
  type: ActivityType;
  title: string;
  detail: string;
  createdAt: string;
}

export interface Dashboard {
  creditBalance: number;
  scansThisHour: number;
  scansLimit: number;
  memberSince: string;
  pairsCirculated: number;
  recentActivity: Activity[];
}

export interface CreditBalance {
  balance: number;
  pending: number;
  lifetimeEarned: number;
  lifetimeSpent: number;
}

export type TradeInUploadMimeType = typeof TradeInUploadMimeType[keyof typeof TradeInUploadMimeType];


export const TradeInUploadMimeType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
} as const;

export interface TradeInUpload {
  /** @minLength 1 */
  imageName: string;
  /**
     * @minLength 1
     * @maxLength 4000000
     */
  imageData: string;
  mimeType: TradeInUploadMimeType;
}

export type TradeInAnalysisStatus = typeof TradeInAnalysisStatus[keyof typeof TradeInAnalysisStatus];


export const TradeInAnalysisStatus = {
  ready: 'ready',
} as const;

export interface TradeInAnalysis {
  id: string;
  status: TradeInAnalysisStatus;
  brand: string;
  model: string;
  condition: string;
  estimatedValue: number;
  creditOffer: number;
  message: string;
}

export type TradeInDecisionInputDecision = typeof TradeInDecisionInputDecision[keyof typeof TradeInDecisionInputDecision];


export const TradeInDecisionInputDecision = {
  accept: 'accept',
  decline: 'decline',
} as const;

export interface TradeInDecisionInput {
  decision: TradeInDecisionInputDecision;
}

export type TradeInSummaryDecision = typeof TradeInSummaryDecision[keyof typeof TradeInSummaryDecision];


export const TradeInSummaryDecision = {
  accepted: 'accepted',
  declined: 'declined',
} as const;

export interface TradeInSummary {
  id: string;
  decision: TradeInSummaryDecision;
  creditAdded: number;
  newBalance: number;
  message: string;
}

export interface CheckoutItemInput {
  productId: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  quantity: number;
}

export interface CheckoutInput {
  /** @minItems 1 */
  items: CheckoutItemInput[];
}

export interface OrderItem {
  productId: string;
  name: string;
  quantity: number;
  unitPrice: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  confirmed: 'confirmed',
} as const;

export interface Order {
  id: string;
  status: OrderStatus;
  total: number;
  creditUsed: number;
  cashDue: number;
  items: OrderItem[];
  createdAt: string;
}