# tradit-edit/lib/api-zod/src/generated/types/checkoutItemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutItemInput {
  productId: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  quantity: number;
}