# tradit-edit/lib/api-zod/src/generated/types/creditBalance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreditBalance {
  balance: number;
  pending: number;
  lifetimeEarned: number;
  lifetimeSpent: number;
}