# tradit-edit/lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  status: OrderStatus;
  total: number;
  creditUsed: number;
  cashDue: number;
  items: OrderItem[];
  createdAt: string;
}