# tradit-edit/lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  brand: string;
  name: string;
  category: string;
  sizeRange: string;
  color: string;
  price: number;
  originalPrice: number;
  imageUrl: string;
  /** @nullable */
  badge?: string | null;
  description: string;
  availableSizes: string[];
}