# tradit-edit/lib/api-zod/src/generated/types/tradeInAnalysis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeInAnalysisStatus } from './tradeInAnalysisStatus';

export interface TradeInAnalysis {
  id: string;
  status: TradeInAnalysisStatus;
  brand: string;
  model: string;
  condition: string;
  estimatedValue: number;
  creditOffer: number;
  message: string;
}