# tradit-edit/lib/api-zod/src/generated/types/tradeInDecisionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeInDecisionInputDecision } from './tradeInDecisionInputDecision';

export interface TradeInDecisionInput {
  decision: TradeInDecisionInputDecision;
}