# tradit-edit/lib/api-zod/src/generated/types/tradeInSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeInSummaryDecision } from './tradeInSummaryDecision';

export interface TradeInSummary {
  id: string;
  decision: TradeInSummaryDecision;
  creditAdded: number;
  newBalance: number;
  message: string;
}