# tradit-edit/lib/api-zod/src/generated/types/tradeInUpload.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeInUploadMimeType } from './tradeInUploadMimeType';

export interface TradeInUpload {
  /** @minLength 1 */
  imageName: string;
  /**
     * @minLength 1
     * @maxLength 4000000
     */
  imageData: string;
  mimeType: TradeInUploadMimeType;
}